requirements.txt
fastapi
uvicorn
motor
pydantic
python-docx
requests
haystack
boto3
ollama-haystack
